Create README.md
# snatchysneakers

This was a project made by me about shoe e-commerce buisness for sneakers called 'Snatchy Sneakers'.
CREDIT (Images) - Yousuf Azeem
Year Made: 2023
Date: Sat, March 4
Grade: 6

This was my first ever project with Javascript + a working fake money system.
This is the final version.

# Fake Currency System (Buy and to Sell)

It has a working fake currency that gets saved automatically and updates every 10 seconds.
The way I fixed the fake money is by having the moneyscript only update and set the default value of money. The way the money variable is saved is in localStorage. 
Each page holds the event listeners.
The pages that show the money variable make a new variable and use that to display the localStorage money variable.